Fix warning of svnapot_test

+++ app/command/svnapot_test.c
@@ -22,7 +22,7 @@ static int cmd_svnapot_test_handler(int argc, char *argv[], void *priv)
 	for (i = 0; i < pnum; i++) {
 		addr[i * PAGE_SIZE] = 66;
 		addr[i * PAGE_SIZE + PAGE_SIZE - 1] = 88;
-		pte = mmu_get_pte(addr + i * PAGE_SIZE);
+		pte = mmu_get_pte((unsigned long)(addr + i * PAGE_SIZE));
 		print("pte:0x%lx\n", *pte);
 		print("addr[%d]:%d addr[%d]:%d\n",
 			i * PAGE_SIZE, addr[i * PAGE_SIZE],
